+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
++ Fixed incorrect default query settings.
+
 ## 1.0.0
 
 Initial release.