+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
+### Improvements
 + Fixed incorrect default query settings.
++ Refactored query function to use pipes instead of nested observables.
 
 ## 1.0.0